reid-service/requirements.txt
# OSNet Person Re-ID Service Requirements - Python 3.11
# Version convergence: FastAPI 0.115.x, Pydantic 2.8+, NumPy 1.26.x, OpenCV 4.10.x
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.0
httpx==0.27.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
onnxruntime-gpu==1.18.1
python-dotenv==1.0.1
python-multipart==0.0.6
prometheus-client==0.21.0
pillow==10.4.0
supabase==2.9.1